Major rewrite of GetTileForRect method, hopefully this fixes a large memory leak. Instead of getting each band separately and copying it into the geotile, we now use the GeoTile->Data() method and the GDAL RasterIO method to directly read the tile data into the GeoTile.
In addition renamed the GetTileSize method to the more aptly named
GetBlockSize. Using GetTileSize to calculate the size of the buffer to
read might have resulted in a memory leak. The block size given back
from this method is the optimum block size to be read from a
GDALDataset. This however doesn't correspond to the tile size that was
set in the pipeline settings; this resulted in more bytes being read
then being copied to the GeoTile.